darwin-compute/core/requirements.txt
setuptools==67.8.0
mysql-connector-python~=9.4.0
PyYAML~=6.0
shortuuid~=1.0.11
python-dateutil~=2.8.2
pytest-mock~=3.12.0
aiohttp~=3.9.5
boto3~=1.34.54
elasticsearch~=7.10.1
typeguard~=2.13.3
requests~=2.28.1
httpx~=0.23.3
async-lru~=2.0.5
opentelemetry-sdk~=1.38.0
opentelemetry-api~=1.38.0
opentelemetry-exporter-otlp-proto-http~=1.38.0
opentelemetry-instrumentation-fastapi~=0.59b0
opentelemetry-instrumentation-requests~=0.59b0
opentelemetry-instrumentation-logging~=0.59b0
opentelemetry-instrumentation-mysql~=0.59b0
opentelemetry-instrumentation-elasticsearch~=0.59b0
